+++ src/main/java/com/alexvr/bedres/capability/abilities/PlayerAbilityStorage.java
@@ -18,7 +18,7 @@ public class PlayerAbilityStorage implements Capability.IStorage<IPlayerAbility>
         tag.putString("sword",instance.getSword());
         tag.putString("hoe",instance.getHoe());
         tag.putString("result",instance.getRitualCraftingResult());
-        tag.putInt("speed",instance.getMiningSpeedBoost());
+        tag.putFloat("speed",instance.getMiningSpeedBoost());
         tag.putInt("ritualTimer",instance.getRitualTimer());
         tag.putInt("totalRitual",instance.getRitualTotalTimer());
         tag.putFloat("gravilty",instance.getGravityMultiplier());
@@ -45,7 +45,7 @@ public class PlayerAbilityStorage implements Capability.IStorage<IPlayerAbility>
         instance.setShovel(((CompoundNBT)nbt).getString("shovel"));
         instance.setSword(((CompoundNBT)nbt).getString("sword"));
         instance.setHoe(((CompoundNBT)nbt).getString("hoe"));
-        instance.setMiningSpeedBoost(((CompoundNBT)nbt).getInt("speed"));
+        instance.setMiningSpeedBoost(((CompoundNBT)nbt).getFloat("speed"));
         instance.setJumpBoost(((CompoundNBT)nbt).getFloat("jump"));
         instance.setGRavityMultiplier(((CompoundNBT)nbt).getFloat("gravilty"));
         instance.setRitualTimer(((CompoundNBT)nbt).getInt("ritualTimer"));
